djpyrus/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
pandas==1.5.2
pyrus==1.1.1
pyrus_api==2.12.0
python-dotenv==0.21.0
requests==2.28.1
celery
djoser